pyproject.toml
[tool.poetry]
name = "netapi"
version = "0.2.2"
description = "Network API library. A programmatic interface to network devices"
authors = ["David Flores <davidflores7_8@hotmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://gitlab.com/the-networkers/netaudithor/netapi"
homepage = "https://gitlab.com/the-networkers/netaudithor/netapi"
keywords = ["network", "library", "automation", "netdevops"]
classifiers = [
        "Programming Language :: Python :: 3.8",
]

[tool.poetry.dependencies]
python = "^3.8.5"
pyeapi = "^0.8.2"
netaddr = "^0.7.19"
pyyaml = "^5.1"
pydantic = "^0.32.0"
pendulum = "^2.0"
bitmath = "^1.3"

[tool.poetry.dev-dependencies]
ipython = "^7.5"
flake8 = "^3.7"
black = {version = "^19.3b0",allows-prereleases = true}
tox = "^3.13"
pytest = "^4.5"
mkdocs = "^1.0"
mkdocs-material = "^4.4"
markdown-include = "^0.5"
pydoc-markdown = { git = "https://github.com/NiklasRosenstein/pydoc-markdown.git", branch = "develop", category="dev" }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
